+++ lib/avromatic/io.rb
@@ -1,0 +1,8 @@
+module Avromatic
+  module IO
+    UNION_MEMBER_INDEX = '__avromatic_member_index'.freeze
+  end
+end
+
+require 'avromatic/io/datum_reader'
+require 'avromatic/io/datum_writer'